+++ source/fitz/string.c
@@ -22,7 +22,7 @@ size_t
 fz_strnlen(const char *s, size_t n)
 {
 	const char *p = memchr(s, 0, n);
-	return p ? p - s : n;
+	return p ? (size_t) (p - s) : n;
 }
 
 int
@@ -735,7 +735,7 @@ static char *twoway_memmem(const unsigned char *h, const unsigned char *z, const
 	/* Search loop */
 	for (;;) {
 		/* If remainder of haystack is shorter than needle, done */
-		if (z-h < l) return 0;
+		if ((size_t)(z-h) < l) return 0;
 
 		/* Check last byte first; advance by shift on mismatch */
 		if (BITOP(byteset, h[l-1], &)) {
