+++ source/html/epub-doc.c
@@ -50,7 +50,7 @@ epub_resolve_link(fz_context *ctx, fz_document *doc_, const char *dest, float *x
 	epub_chapter *ch;
 
 	const char *s = strchr(dest, '#');
-	size_t n = s ? s - dest : strlen(dest);
+	size_t n = s ? (size_t)(s - dest) : strlen(dest);
 	if (s && s[1] == 0)
 		s = NULL;
